+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".DrawerActivity" />
+        <activity android:name=".Listen"></activity>
     </application>
 
 </manifest>
