+++ app/src/main/java/com/lonely/solotalk/DrawerActivity.java
@@ -1,0 +1,90 @@
+package com.lonely.solotalk;
+
+import android.app.Activity;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
+import android.view.View;
+import android.widget.AdapterView;
+import android.widget.FrameLayout;
+import android.widget.RelativeLayout;
+
+import com.mikepenz.materialdrawer.Drawer;
+import com.mikepenz.materialdrawer.DrawerBuilder;
+import com.mikepenz.materialdrawer.accountswitcher.AccountHeaderBuilder;
+import com.mikepenz.materialdrawer.model.SecondaryDrawerItem;
+import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem;
+import com.mikepenz.materialdrawer.model.interfaces.Nameable;
+
+
+public class DrawerActivity extends AppCompatActivity {
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        super.setContentView(R.layout.activity_drawer);
+        final Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
+//        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+//        getSupportActionBar().setHomeButtonEnabled(true);
+
+        DrawerBuilder builder = new DrawerBuilder()
+                .withActivity(this)
+                .withToolbar(toolbar)
+                .addDrawerItems(
+                        new SecondaryDrawerItem().withIcon(R.drawable.drawer).withName(R.string.drawer_title1)
+
+                )
+                .withOnDrawerItemClickListener(new Drawer.OnDrawerItemClickListener() {
+                    @Override
+                    public boolean onItemClick(AdapterView<?> parent, View view, int position, long id, IDrawerItem drawerItem) {
+
+                        if (drawerItem != null) {
+                            if (drawerItem instanceof Nameable) {
+                                toolbar.setTitle(((Nameable)
+                                        drawerItem).getNameRes());
+                            }
+                        }
+
+                        return false;
+
+                    }
+                });
+
+        AccountHeader headerResult = new AccountHeaderBuilder()
+                .withActivity(this)
+                .withHeaderBackground(R.drawable.header)
+                .addProfiles(
+                        new ProfileDrawerItem().withName("Mike Penz").withEmail("mikepenz@gmail.com").withIcon(getResources().getDrawable(R.drawable.profile))
+                )
+                .withOnAccountHeaderListener(new AccountHeader.OnAccountHeaderListener() {
+                    @Override
+                    public boolean onProfileChanged(View view, IProfile profile, boolean currentProfile) {
+                        return false;
+                    }
+                })
+                .build();
+
+        Drawer drawer = builder.build();
+
+        if (savedInstanceState == null) {
+            // on first time display view for first nav item
+            return;
+        }
+
+        //fullLayout = (RelativeLayout) getLayoutInflater().inflate(R.layout.activity_drawer, null);
+        //frameLayout = (FrameLayout) fullLayout.findViewById(R.id.drawer_frame);
+
+        //getLayoutInflater().inflate(layoutResID, frameLayout, true);
+
+
+        //Your DrawerActivity content...
+
+    }
+
+
+}
+
