+++ app/src/main/java/com/lonely/solotalk/Listen.java
@@ -1,0 +1,13 @@
+package com.lonely.solotalk;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class Listen extends DrawerActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_listen);
+    }
+}
