+++ app/src/main/res/drawable-mdpi/settings.png
