+++ app/src/main/res/layout/activity_home_page.xml
@@ -1,26 +1,22 @@
-<?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/activity_home_page"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.lonely.solotalk.HomePage">
+    tools:context=".HomePage">
 
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#BD601B"
-        android:elevation="4dp"
-        android:minHeight="?attr/actionBarSize"
-        android:transitionName="actionBar"
-        style="@style/HeaderBar"
-        app:theme="@style/ActionBarThemeOverlay"
-        app:popupTheme="@style/ActionBarPopupThemeOverlay"
-        />
+
+    <!--<android.support.v7.widget.Toolbar-->
+        <!--android:id="@+id/toolbar"-->
+        <!--android:layout_width="match_parent"-->
+        <!--android:layout_height="wrap_content"-->
+        <!--android:background="#BD601B"-->
+        <!--android:elevation="4dp"-->
+        <!--android:minHeight="?attr/actionBarSize"-->
+        <!--android:transitionName="actionBar"-->
+        <!--style="@style/HeaderBar"-->
+        <!--app:theme="@style/ActionBarThemeOverlay"-->
+        <!--app:popupTheme="@style/ActionBarPopupThemeOverlay"-->
+        <!--/>-->
+
 </RelativeLayout>
